Added base class shared between client and server api implementation.
